requirements.txt
Flask
openai
flask-cors
pyautogen